Address PR reviews
